requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==2.2.8
django-jet==1.0.8
django-celery-beat==2.2.0
django-colorfield==0.4.1
django-colorful==1.3
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==2.4.0
django-mathfilters==1.0.0
django-phonenumber-field==5.0.0
django-resized==0.3.11
django-templateaddons==0.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
gunicorn==20.0.4
importlib-metadata==4.0.0
kombu==5.0.2
phonenumbers==8.12.18
Pillow==8.0.1
prompt-toolkit==3.0.16
psycopg2-binary==2.8.6
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==3.7.4.3
Unidecode==1.2.0
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.1
uWSGI
